Modal Implementation, Query additions and more user information
