+++ packages/handler/src/interactions/commands/user.ts
@@ -3,7 +3,8 @@ import { fetchAnilistUser } from "database";
 import { ApplicationCommandOptionType } from "discord-api-types/v10";
 import { Logger } from "log";
 import type { Command } from "../../services/commands.js";
-import { fetchAnilistUserData, getCommandOption, intervalTime } from "../../utility/interactionUtils.js";
+import { fetchAnilistUserData, intervalTime } from "../../utility/anilistUtil.js";
+import { getCommandOption } from "../../utility/interactionUtils.js";
 
 const logger = new Logger();
 export const interaction: Command = {
@@ -16,25 +17,21 @@ export const interaction: Command = {
         let username = getCommandOption("username", ApplicationCommandOptionType.String, interaction.options);
 
         if (username === null) {
-            username = (await fetchAnilistUser(interaction.member_id)).username;
-
-            logger.debug(`Attempt fetched user from database: ${username}`, "User");
-            if (username === null) {
-                logger.warn("No user found in database", "User");
-                return interaction.edit({
-                    content: "Please provide a username or link your Anilist account",
-                    ephemeral: true,
-                });
+            logger.debug("Attemping fetching user from database", "User");
+            try {
+                username = (await fetchAnilistUser(interaction.member_id)).username;
+            } catch (error: any) {
+                logger.error(`Error fetching user from database: ${error}`, "User");
+                return interaction.reply({ content: "Please setup your account with /setup!", ephemeral: true });
             }
         }
 
         logger.debug(`Fetching user: ${username}`, "User");
         const userFetch = await fetchAnilistUserData(username, interaction);
         if (userFetch === null) {
-            return interaction.edit({ content: "User not found", ephemeral: true });
+            return interaction.reply({ content: "User not found", ephemeral: true });
         }
 
-        logger.debug("Gained user data", "User", userFetch.result);
         const footer = `${userFetch.result.dataFrom === "API" ? "Data from Anilist API" : `Displaying cached data : refreshes in ${intervalTime(userFetch.result.leftUntilExpire)}`}`;
         const embed = new EmbedBuilder()
             .setTitle(userFetch.result.name)
